// types/MovieType.tsx
export interface MovieType {
    id: string;
    created_at: number;

    title_uz: string;
    title_ru: string;
    title_en: string;

    description_uz: string;
    description_ru: string;
    description_en: string;

    poster_url: string;
    banner_url: string;
    trailer_url: string;
    video_url: string;

    duration_minutes: number;
    release_year: number;
    imdb_rating: number;

    age_rating: string;
    country: string;
    language: string;

    is_premium: boolean;
    is_featured: boolean;
    is_active: boolean;

    view_count: number;

    updated_at: number;
    created_by: number;
  };